+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - [TD-3894] User email is now optional
 - Removed deprecated `is_admin` field in user requests and responses
 
+### Added 
+
+- [TD-3617] url in state for auth0 and idc
+
 ## [4.23.0] 2021-06-28
 
 ### Added